Move CorsHelper to internal.utils.CorsUtil
It is an implementation detail (no external API surface), so it
belongs under internal/utils alongside KinoticUtil/EventUtil rather
than in the public api/config package.
